requirements.txt
Django>=4.2.21
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
whitenoise>=6.6.0
python-decouple>=3.8
